Move markdown fns into ssgs
