--- test-files/jekyll/commonmark.yaml
@@ -1,11 +1,0 @@
-markdown: CommonMark
-commonmark:
-  options:
-    - HARDBREAKS
-    - GFM_QUIRKS
-  extensions:
-    - strikethrough
-    - table
-    - autolink
-    - superscript
-    - header_ids
